backup de extraccion de datos/requirements.txt
telethon
reportlab
python-dotenv
requests
supabase